BFS route
Walker can use BFS routes for routing.

+++ src/main/java/ru/drifles/crpg/common/Position.java
@@ -3,41 +3,41 @@ package ru.drifles.crpg.common;
 import java.util.Objects;
 
 public final class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Position() {
         this.x = 0;
         this.y = 0;
     }
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public Position(int xy) {
+    public Position(double xy) {
         this.x = xy;
         this.y = xy;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void setXY(int x, int y) {
+    public void setXY(double x, double y) {
         this.x = x;
         this.y = y;
     }
